# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register with email, phone & password",
        "description": "Does not return tokens. Sends verification email. Use POST /auth/verify-email to activate.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterPendingResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login with email & password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Email not verified — verification email resent"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "summary": "Verify email with OTP code",
        "description": "Returns JWT tokens after successful verification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "summary": "Resend email verification code",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/phone/send-otp": {
      "post": {
        "operationId": "AuthController_sendPhoneOtp",
        "summary": "Send OTP to phone number",
        "description": "Rate limited. OTP expires in 5 minutes.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPhoneOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/phone/verify": {
      "post": {
        "operationId": "AuthController_verifyPhoneOtp",
        "summary": "Verify phone OTP",
        "description": "Auto-registers or logs in user. No password required.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPhoneOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "summary": "Request password reset OTP",
        "description": "Send OTP via email or phone. Rate limited.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-reset-otp": {
      "post": {
        "operationId": "AuthController_verifyResetOtp",
        "summary": "Verify password reset OTP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyResetOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "summary": "Reset password",
        "description": "Invalidates all refresh tokens after reset.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid OTP or user not found"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "summary": "Refresh access token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked refresh token"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "summary": "Get current user profile and permissions",
        "description": "Use on app load to restore session. Store `permissions` in localStorage.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/categories": {
      "get": {
        "operationId": "CategoryController_findAll",
        "summary": "List category tree (paginated main categories + subcategories)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max 100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/{slug}": {
      "get": {
        "operationId": "CategoryController_findBySlug",
        "summary": "Get category by slug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "apartment",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Aqar API",
    "description": "Rental Property Platform — Authentication, Authorization (RBAC), Properties & Bookings",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ahmed Ali"
          },
          "email": {
            "type": "string",
            "example": "ahmed@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+201234567890"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "OWNER"
            ],
            "example": "CUSTOMER",
            "description": "CUSTOMER or OWNER only"
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "password",
          "role"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-here"
          },
          "name": {
            "type": "string",
            "example": "Ahmed Ali"
          },
          "email": {
            "type": "object",
            "example": "ahmed@example.com",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "example": "+201234567890",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "OWNER",
              "CUSTOMER"
            ],
            "example": "CUSTOMER"
          },
          "isVerified": {
            "type": "boolean",
            "example": false,
            "description": "Whether email has been verified"
          },
          "isProfileComplete": {
            "type": "boolean",
            "description": "True if owner submitted extended profile (not INCOMPLETE)",
            "example": false
          },
          "profileStatus": {
            "type": "string",
            "enum": [
              "INCOMPLETE",
              "BASIC_DONE",
              "KYC_PENDING",
              "VERIFIED",
              "REJECTED"
            ],
            "example": "INCOMPLETE"
          },
          "ownerType": {
            "type": "object",
            "example": null,
            "description": "Set after profile completion"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "role",
          "isVerified"
        ]
      },
      "RegisterPendingResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Registration successful. Please verify your email."
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "message",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ahmed@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token (15 min)"
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token (7 days)"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "permissions": {
            "example": [
              "property.read",
              "booking.create"
            ],
            "description": "Permission actions for the user role (store in localStorage)",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user",
          "permissions"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ahmed@example.com"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "6-digit code from verification email"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ahmed@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OTP sent successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "SendPhoneOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+201234567890"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyPhoneOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+201234567890"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "6-digit OTP from SMS"
          },
          "name": {
            "type": "string",
            "example": "Ahmed Ali",
            "description": "Used when auto-registering a new user"
          }
        },
        "required": [
          "phone",
          "code",
          "name"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ahmed@example.com",
            "description": "Required if phone is not provided"
          },
          "phone": {
            "type": "string",
            "example": "+201234567890",
            "description": "Required if email is not provided"
          }
        }
      },
      "VerifyResetOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ahmed@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+201234567890"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "6-digit OTP code"
          }
        },
        "required": [
          "code"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ahmed@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+201234567890"
          },
          "code": {
            "type": "string",
            "example": "123456"
          },
          "newPassword": {
            "type": "string",
            "example": "newPassword123",
            "minLength": 8
          }
        },
        "required": [
          "code",
          "newPassword"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token from login/register response"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "permissions": {
            "example": [
              "property.read",
              "booking.create"
            ],
            "description": "Permission actions for the user role",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "user",
          "permissions"
        ]
      },
...
    }
  }
}